+++ src/components/input/script.ts
@@ -72,6 +72,9 @@ export default class StInput extends Vue implements ValidatableComponent<string>
   @Prop(Number)
   tabindex!: number;
 
+  @Prop(Boolean)
+  preventInput?: boolean;
+
   @Prop(ComponentValidator)
   validator?: ComponentValidator<string>;
 
@@ -137,7 +140,9 @@ export default class StInput extends Vue implements ValidatableComponent<string>
   handleInput(event: Event) {
     const { value } = event.target as HTMLInputElement;
     this.$emit('input', value);
-    this.setInputValue(value);
+    if (!this.preventInput) {
+      this.setInputValue(value);
+    }
   }
 
   handleChange(event: Event) {
@@ -173,12 +178,16 @@ export default class StInput extends Vue implements ValidatableComponent<string>
     this.$emit('change', '');
     this.setInputValue('');
     if (this.focusAfterClear) {
-      this.focusInput();
+      this.focus();
     }
     this.$emit('clear');
   }
 
-  focusInput() {
+  focus() {
     (this.$refs.input as HTMLInputElement).focus();
+  }
+
+  blur() {
+    (this.$refs.input as HTMLInputElement).blur();
   }
 }