+++ src/components/select/_select-single/script.ts
@@ -8,7 +8,10 @@ import StSelectDropdown from '../_select-dropdown/index.vue';
 import StSelectDropdownScript from '../_select-dropdown/script';
 import StSelectBase from '../_select.base';
 import StIcon from '../../icon/index.vue';
-import { SelectOption } from '../types';
+import {
+  SelectOption,
+  SingleSelectValue,
+} from '../types';
 
 
 @Component({
@@ -21,7 +24,7 @@ import { SelectOption } from '../types';
 })
 export default class StSelectSingle extends StSelectBase {
   @Prop(String)
-  value!: string;
+  value!: SingleSelectValue;
 
   dropdownVisible: boolean = false;
 
@@ -31,7 +34,6 @@ export default class StSelectSingle extends StSelectBase {
       selected: this.selectedOption && this.selectedOption.value === option.value,
     }));
   }
-
 
   get selectedOption(): SelectOption | undefined {
     return this.options.find((option: SelectOption) => option && option.value === this.value);