+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 0.2.14
+
+### Features
+* Added `prevent-input` prop to `Input`.
+* Added `header-style` prop to Column of `Table` component for styling header cell in table
+* Added `focus` and `blur` methods to `Input` and `Autocomplete`
+* Added `useReferenceWidth` option to storybook of `Autocomplete`
+
+### Maintenance
+* Fixed weird bug when typing in Select leaded to displaying clear icon
+* Fixed `Select` arrow visibility when component is non-clearable
+* Fixed bug with `useReferenceWidth` do not work properly in `Dropdown`
+* Fixed checkbox height (container of checkbox's square was a bit higher)
+* Fixed multiple `Select`'s selected options update. There was a problem with displaying of collapser's items when `Select`'s value was changed from outside
+* Created `BaseSelectValue`, `SingleSelectValue`, and `MultipleSelectValue` types for `Select` prop `value`   
+
+
+
+
 ## 0.2.13
 
 ### Features